ENG-4585: Apply minor suggestions
